Changed file structure, deleted UI interface and mainGUI because they were for no use, added packages
