+++ nl.avans.netflixstatistix/src/data/ITable.java
@@ -1,0 +1,4 @@
+package data;
+
+public interface ITable {
+}
