+++ nl.avans.netflixstatistix/src/data/Table.java
@@ -1,0 +1,5 @@
+package data;
+
+public enum Table {
+    PROGRAM, FILM, SERIE, SEASON, EPISODE, ACCOUNT, PROFILE
+}
