--- nl.avans.netflixstatistix/src/main/java/data/ITable.java
@@ -1,4 +1,0 @@
-package main.java.data;
-
-public interface ITable {
-}
