--- nl.avans.netflixstatistix/src/main/java/data/Table.java
@@ -1,5 +1,0 @@
-package main.java.data;
-
-public enum Table {
-    PROGRAM, FILM, SERIE, SEASON, EPISODE, ACCOUNT, PROFILE
-}
