--- nl.avans.netflixstatistix/src/main/java/presentation/view/UI.java
@@ -1,7 +1,0 @@
-package main.java.presentation.view;
-
-import java.awt.*;
-
-public interface UI {
-    public void createComponents(Container container);
-}
