--- nl.avans.netflixstatistix/src/presentation/MainGUI.java
@@ -1,4 +1,0 @@
-package presentation;
-
-public class MainGUI {
-}
